% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_lotka_volterra.R
\docType{data}
\name{LV_DEFAULTS}
\alias{LV_DEFAULTS}
\title{Default Lotka-Volterra parameters}
\format{
An object of class \code{list} of length 4.
}
\usage{
LV_DEFAULTS
}
\description{
\itemize{
\item \code{alpha}: prey birth rate (per capita, per time unit)
\item \code{beta}: predation rate (prey removed per predator per prey per time)
\item \code{gamma}: predator death rate (per capita)
\item \code{delta}: predator birth rate per prey eaten
}
}
\details{
Chosen to produce approximately 3-4 oscillation cycles within the
default time window of 200 time units.
}
\keyword{datasets}
