% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\docType{data}
\name{SLIDE_ACCENT}
\alias{SLIDE_ACCENT}
\title{Slide accent colour (amber, used for frame titles and block borders)}
\format{
An object of class \code{character} of length 1.
}
\usage{
SLIDE_ACCENT
}
\description{
Slide accent colour (amber, used for frame titles and block borders)
}
\keyword{datasets}
