% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\docType{data}
\name{SLIDE_BG}
\alias{SLIDE_BG}
\title{Slide background colour (dark navy)}
\format{
An object of class \code{character} of length 1.
}
\usage{
SLIDE_BG
}
\description{
Slide background colour (dark navy)
}
\keyword{datasets}
