% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\docType{data}
\name{SLIDE_PANEL}
\alias{SLIDE_PANEL}
\title{Slide code-panel background (slightly lighter navy)}
\format{
An object of class \code{character} of length 1.
}
\usage{
SLIDE_PANEL
}
\description{
Slide code-panel background (slightly lighter navy)
}
\keyword{datasets}
