% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\docType{data}
\name{SLIDE_TEXT}
\alias{SLIDE_TEXT}
\title{Slide body text colour (off-white)}
\format{
An object of class \code{character} of length 1.
}
\usage{
SLIDE_TEXT
}
\description{
Slide body text colour (off-white)
}
\keyword{datasets}
